+++ src/database/migrations/2003_01_01_000030_comunik_create_table_contacts_groups.php
@@ -11,18 +11,21 @@ class ComunikCreateTableContactsGroups extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('005_042_contacts_groups', function($table){
-			$table->engine = 'InnoDB';
-			$table->integer('group_042')->unsigned();
-			$table->integer('contact_042')->unsigned();
+		if (! Schema::hasTable('005_042_contacts_groups'))
+		{
+			Schema::create('005_042_contacts_groups', function ($table) {
+				$table->engine = 'InnoDB';
+				$table->integer('group_042')->unsigned();
+				$table->integer('contact_042')->unsigned();
 
-			$table->primary(['contact_042', 'group_042']);
+				$table->primary(['contact_042', 'group_042']);
 
-			$table->foreign('group_042', 'fk01_005_042_contacts_groups')->references('id_040')->on('005_040_group')
+				$table->foreign('group_042', 'fk01_005_042_contacts_groups')->references('id_040')->on('005_040_group')
 					->onDelete('cascade')->onUpdate('cascade');
-			$table->foreign('contact_042', 'fk02_005_042_contacts_groups')->references('id_041')->on('005_041_contact')
-				->onDelete('cascade')->onUpdate('cascade');
-		});
+				$table->foreign('contact_042', 'fk02_005_042_contacts_groups')->references('id_041')->on('005_041_contact')
+					->onDelete('cascade')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
