+++ src/database/migrations/2003_01_01_000060_comunik_create_table_email_campaigns_countries.php
@@ -12,18 +12,21 @@ class ComunikCreateTableEmailCampaignsCountries extends Migration {
 	 */
 	public function up()
 	{
-        Schema::create('005_045_email_campaigns_countries', function(Blueprint $table){
-            $table->engine = 'InnoDB';
-            $table->integer('campaign_045')->unsigned();
-            $table->string('country_045', 2);
+		if (! Schema::hasTable('005_045_email_campaigns_countries'))
+		{
+			Schema::create('005_045_email_campaigns_countries', function (Blueprint $table) {
+				$table->engine = 'InnoDB';
+				$table->integer('campaign_045')->unsigned();
+				$table->string('country_045', 2);
 
-            $table->primary(['campaign_045', 'country_045'], 'pk01_005_045_email_campaigns_countries');
+				$table->primary(['campaign_045', 'country_045'], 'pk01_005_045_email_campaigns_countries');
 
-            $table->foreign('campaign_045', 'fk01_005_045_email_campaigns_countries')->references('id_044')->on('005_044_email_campaign')
-                    ->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('country_045', 'fk02_005_045_email_campaigns_countries')->references('id_002')->on('001_002_country')
-                    ->onDelete('cascade')->onUpdate('cascade');
-        });
+				$table->foreign('campaign_045', 'fk01_005_045_email_campaigns_countries')->references('id_044')->on('005_044_email_campaign')
+					->onDelete('cascade')->onUpdate('cascade');
+				$table->foreign('country_045', 'fk02_005_045_email_campaigns_countries')->references('id_002')->on('001_002_country')
+					->onDelete('cascade')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
