+++ src/database/migrations/2003_01_01_000070_comunik_create_table_email_campaigns_groups.php
@@ -12,18 +12,21 @@ class ComunikCreateTableEmailCampaignsGroups extends Migration {
 	 */
 	public function up()
 	{
-        Schema::create('005_046_email_campaigns_groups', function(Blueprint $table){
-            $table->engine = 'InnoDB';
-            $table->integer('campaign_046')->unsigned();
-            $table->integer('group_046')->unsigned();
+		if (! Schema::hasTable('005_046_email_campaigns_groups'))
+		{
+			Schema::create('005_046_email_campaigns_groups', function (Blueprint $table) {
+				$table->engine = 'InnoDB';
+				$table->integer('campaign_046')->unsigned();
+				$table->integer('group_046')->unsigned();
 
-            $table->primary(['campaign_046', 'group_046'], 'pk01_005_046_email_campaigns_groups');
+				$table->primary(['campaign_046', 'group_046'], 'pk01_005_046_email_campaigns_groups');
 
-            $table->foreign('campaign_046', 'fk01_005_046_email_campaigns_groups')->references('id_044')->on('005_044_email_campaign')
-                    ->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('group_046', 'fk02_005_046_email_campaigns_groups')->references('id_040')->on('005_040_group')
-                    ->onDelete('cascade')->onUpdate('cascade');
-        });
+				$table->foreign('campaign_046', 'fk01_005_046_email_campaigns_groups')->references('id_044')->on('005_044_email_campaign')
+					->onDelete('cascade')->onUpdate('cascade');
+				$table->foreign('group_046', 'fk02_005_046_email_campaigns_groups')->references('id_040')->on('005_040_group')
+					->onDelete('cascade')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
