+++ src/database/migrations/2003_01_01_000090_comunik_create_table_email_send_historical.php
@@ -12,27 +12,30 @@ class ComunikCreateTableEmailSendHistorical extends Migration {
 	 */
 	public function up()
 	{
-        Schema::create('005_048_email_send_historical', function(Blueprint $table){
-            $table->engine = 'InnoDB';
-            $table->bigIncrements('id_048')->unsigned();
-            $table->bigInteger('send_queue_048')->nullable()->unsigned();
-            $table->integer('campaign_048')->unsigned();
-            $table->integer('contact_048')->unsigned();
+        if (! Schema::hasTable('005_048_email_send_historical'))
+        {
+            Schema::create('005_048_email_send_historical', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->bigIncrements('id_048')->unsigned();
+                $table->bigInteger('send_queue_048')->nullable()->unsigned();
+                $table->integer('campaign_048')->unsigned();
+                $table->integer('contact_048')->unsigned();
 
-            // date when he was created
-            $table->integer('create_048')->unsigned();
-            // date when he was sent
-            $table->integer('sent_048')->unsigned();
-            // count number times has been seen
-            $table->integer('viewed_048')->unsigned()->default(0);
+                // date when he was created
+                $table->integer('create_048')->unsigned();
+                // date when he was sent
+                $table->integer('sent_048')->unsigned();
+                // count number times has been seen
+                $table->integer('viewed_048')->unsigned()->default(0);
 
-            $table->foreign('send_queue_048', 'fk01_005_048_email_send_historical')->references('id_047')->on('005_047_email_send_queue')
-                ->onDelete('set null')->onUpdate('cascade');
-            $table->foreign('campaign_048', 'fk02_005_048_email_send_historical')->references('id_044')->on('005_044_email_campaign')
-                ->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('contact_048', 'fk03_005_048_email_send_historical')->references('id_041')->on('005_041_contact')
-                ->onDelete('cascade')->onUpdate('cascade');
-        });
+                $table->foreign('send_queue_048', 'fk01_005_048_email_send_historical')->references('id_047')->on('005_047_email_send_queue')
+                    ->onDelete('set null')->onUpdate('cascade');
+                $table->foreign('campaign_048', 'fk02_005_048_email_send_historical')->references('id_044')->on('005_044_email_campaign')
+                    ->onDelete('cascade')->onUpdate('cascade');
+                $table->foreign('contact_048', 'fk03_005_048_email_send_historical')->references('id_041')->on('005_041_contact')
+                    ->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
 	}
 
 	/**
